[fix] small fixes and subtle refactoring for the testing setup
