+++ lib/construct/index.js
@@ -109,6 +109,7 @@ class Constructor extends EventEmitter {
    * @api public
    */
   build({ promote, data }, done) {
+    const { app } = this;
     const progress = new Progress({
       nsq: {
         writer: this.nsq.writer,
@@ -116,7 +117,6 @@ class Constructor extends EventEmitter {
         log: app.contextLog
       }
     }, done);
-    const { app } = this;
 
     //
     // Compile build specifications from the data.
