+++ lib/construct/status-writer.js
@@ -1,6 +1,5 @@
 const { performance } = require('perf_hooks');
-const { Writable } = require('stream');
-const retry = require('retryme');
+const nsqStream = require('./nsq-stream');
 
 const eventTypes = {
   error: 'error',
@@ -26,7 +25,7 @@ class StatusWriter {
   constructor(opts) {
     const { nsq = {} } = opts || {};
     this.log = opts.log || { info: function () {}, error: function () {} };
-    this.nsqStream = nsq.writer && nsq.topic && new NsqRetryWriteStream(nsq);
+    this.nsqStream = nsq.writer && nsq.topic && nsqStream.createWriteStream(nsq.writer, nsq.topic);
     if (this.nsqStream) this.nsqStream.on('error', this._connectionError.bind(this));
     this.metadata = opts && opts.metadata || {};
     this.buildsCompleted = 0;
@@ -204,34 +203,3 @@ class StatusWriter {
 
 module.exports = StatusWriter;
 
-/**
- * @class NsqRetryWriteStream
- */
-class NsqRetryWriteStream extends Writable {
-  /**
-   * @param {Object} options For the stream
-   *  @param {String} options.topic Topic to publish to on nsq
-   *  @param {Writer} options.writer NsqWriter to use for publishing
-   *  @param {Ojbect} options.retryOpts RetryOptions configured for this;
-   * @constructor
-   */
-  constructor({ topic, writer, retryOpts = { retries: Infinity, min: 300, max: 5000 }, ...opts }) {
-    super({ objectMode: true, ...opts });
-    this.topic = topic;
-    this.writer = writer;
-    this.retryOpts = retryOpts;
-  }
-  /**
-   * _write function to do the retryable publish for the nsq writer via the
-   * stream interface
-   *
-   * @param {Object} data Payload to be published to nsq
-   * @param {String} enc Useless encoding meant for other stream types
-   * @param {Function} cb Continuation
-   */
-  _write(data, enc, cb) {
-    retry.op(this.retryOpts).attempt((next) => {
-      this.writer.publish(data, next);
-    }, cb)
-  }
-}
